benchmark/requirements.txt
resource
numpy
pandas
pyyaml
pymongo
sqlalchemy
psycopg2
